Fix armatures not unlinking from scene collection
Turns out the objects list includes *all* objects in a collection, so that means armatures too. This resulted in armatures that happened to be in a collection with meshes for the model to be linked twice but only unlinked once. Now I cast `armatures + objects` to a set to remove duplicates first; problem solved.