Create addon preferences and move games there
The upside is not having to set it in every blend file (which was mostly a problem when collaborating).
The downside is that you lose all your games settings when you disable the addon.
Perhaps I'll back that stuff up to a json and load it on registery. Maybe later.