Calculate tangents for displacements
These are not based on UV's, but just the normal and one of the edges of the face.
It's just so they don't look so hideous in hammer; the user is meant to do texture alignment manually anyway so it doesn't matter that they're upside down sometimes.
I don't think I can support vertex colors btw because you can only paint on the bottom layer of the multires, so that's another thing you'll have to do in hammer.
I'll do material though because eh I did it for brushes so I might as well. However instead of setting faces that don't have material to nodraw, which I don't think makes sense for displacements, I'll set it to `dev/dev_blendmeasure`.