Add align to grid support to displacement export
This required moving some stuff around, and making some new functions.
The subd object now has transforms applied to the mesh itself, before the modifier is added and the final bmesh for it is created.
Also I'm making the base bmesh at the end instead of the start now for aesthetic reasons I guess.
One thing to note is that numbers have this `e-#` thing if they're really small; maybe I should fix that.