Add backup and restore buttons to addon prefs UI
And there we are.
Another thing I'd like to look at is why addon prefs don't save to disk when I edit the games list.
I'm guessing it's because it's not in the preferences UI?
Maybe I can call `bpy.ops.wm.save_userpref()` after every change made to a game or the game's list as a whole, I dunno.