Clean up new attachment code
Very clever solution using prop_search :+1:
For the panel I added some newlines for clarity, and changed double quotes to single quotes.
For the property group I removed the extra dot and double quote at the end of the tooltips.
For the QC generator I removed some brackets I deemed unnecessary, and made the way strings are written a little easier to comprehend.
It was opening a double quote on one line, then closing it on another, no thanks.
Now the lines are a little longer, but more consistent with the rest of the code, with a space at the start of each string.