Add pose bone transforms to right click menu
This is a little operator I made for Axale on the SourceOps discord.
Its purpose is to get values for $proceduralbones.
Rather than putting the buttons in a panel, I'm adding them to the pose mode right click menu.
I'm also giving the ops module init file a long overdue make over to use a tuple of classes.