Use FIXED and CLIP for triangulation
I believe this is what Blender does internally, in order to draw these shapes (it's cheaper / easier to only draw triangles).
Exporting an SMD with SourceOps and then importing it with BST, I can't see a difference now, even with the zebra matcap.
Except with the nurbs surface, which is just slightly different.
But oh well, people only really uses meshes anyway.