Swap the order of checks in panel if statements
This makes more sense I think, especially for the places where it was checking the panel first, and then the existence of the thing that owns the panel enum.
Though there's no way that would ever not be there.. it's just so the checks are consistent.
Doesn't really matter, don't worry about it.