Add prepend_armature and ignore_transforms to FBX
I was worried I wouldn't be able to make these work, but it was not that hard.
When you rename a bone in Blender, all vertex groups that depend on it are automatically renamed too.
I tried setting matrix_world to Identity, but that had problems, so now I'm doing it with matrix_local and matrix_parent_inverse; haven't run into issues with this approach yet.
To make export_scene.fbx behave properly, I had to get rid of leaf bones and animation (which is fine, cause they weren't used anyway).