Fix origin and rotation code in QC generator
So basically, Source 1 is a dumpster fire.
Apparently rotation should be negated, and for SMD it needs to be rotated -90 degrees (the [VDC about SMD](https://developer.valvesoftware.com/wiki/Studiomdl_Data) says so actually).
What surprised me is that for FBX, apparently origin X and Y need to be swapped, one of them made negative, and rotation offset by 180 degrees. I learned this through trial and error.
Also got rid of the minus for Z set through UI (I mistakenly told Cabbage to add it, whoops), and some spaces that didn't need to be on an empty line.