Bump version to 0.5.6
This version gives you the option to use an object as origin.
Also it fixes the issues with orientation and location that SMD and FBX had, so don't be surprised if things are aligned differently now, they are just aligned correctly for the first time.
Oh and I changed the list of authors to be in alphabetical order (aside from my name at the start cause I'm the head honcho).