Properly restore armature after exporting action
If the armature has no animation data, it will be added, and later removed.
If the armature has no action, a backup of its pose will be made, and later restored.
I thought it made sense to backup the pose regardless, since you might have a pose that doesn't conform to your selected action, if it's WIP or something; but it doesn't seem to want to restore the pose if there's an action selected, so I don't bother.
At least this way the code does what it says it does, instead of saying one thing and doing another, if that makes sense.