Clear animation data before FBX export
If there's an armature, and it has animation data, and this animation data has an action, then keep a reference to that action, clear the animation data, and and reset all the pose bones to their default transforms.
After export, if there is an action to restore, new animation data is created for the armature, and the action is set.