Add a bunch of stuff to the README
Now it's actually kind of useful.
I still need to add a text/image based guide to the Usage section, which will be a lot of work.
For now it just links to an old tutorial I made, which I'll replace with a new video when I get the time to record one.