Force static when there is no armature
Instead of refusing to generate the QC, it now just uses $staticprop when there is no armature.
The user is informed of this in parentheses; hopefully that's good enough to avoid confusion.

This is something that messed up a lot of new users, because what people do when confronted with an error message they don't understand, is whatever it takes to make it shut up.
In practice this usually meant selecting any object as armature, and then going to the discord to ask why the mesh they selected wasn't a valid armature.