Only switch to object mode when necessary
Instead of always setting it to object mode, it first checks that we aren't already in object mode.
The reason for this is trying to switch to object mode, even if you're already there, causes an error when the active object is not editable (for example a linked library).
I don't think a non editable object can be in anything but object mode, so this should fix it.
I'm doing this for SMD and VMF export, not messing with rig simulation.
There's something weird about the VMF configure scene code compared to the SMD version, but I'm not going to get sidetracked by that right now.